requirements.txt
pandas
requests
tweepy
gspread
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
py-cord
mailersend
atproto
flask
pytest
